Remove config as argument for usage
When the dupsifter config was used as an argument for the usage
function, any values that had been changed would then be used as the
"default" values for the usage output. However, this is almost always
incorrect. By now initializing a fresh config in the usage function, the
actual default values will be used. It also makes keeping track of where
all the default values are stored easier by keeping them in one init
function.